discordでのログイン機能追加, ページ全体の保護機能追加 #8
