+++ src/app/archive/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function archive() {
+  return (
+    <div>
+      <h1>ここはアーカイブ見る場所</h1>
+    </div>
+  )
+}
