+++ src/app/middleware.ts
@@ -1,0 +1,3 @@
+// アプリケーション全体のページを保護する
+// 参考 => https://next-auth.js.org/configuration/nextjs#middleware
+export { default } from 'next-auth/middleware'
